hexagonal architecture

+++ src/main/java/com/example/hexagonalarchitecturefeatures/core/domain/Product.java
@@ -9,11 +9,11 @@ public class Product {
     private String category;
     private Double price;
     private Integer stock;
-    private Integer cost;
+    private Double cost;
     private LocalDateTime creationDate;
 
 
-    public Product(Long id, String name, String category, Double price, Integer stock, Integer cost, LocalDateTime creationDate) {
+    public Product(Long id, String name, String category, Double price, Integer stock, Double cost, LocalDateTime creationDate) {
         this.id = id;
         this.name = name;
         this.category = category;
@@ -63,11 +63,11 @@ public class Product {
         this.stock = stock;
     }
 
-    public Integer getCost() {
+    public Double getCost() {
         return cost;
     }
 
-    public void setCost(Integer cost) {
+    public void setCost(Double cost) {
         this.cost = cost;
     }
